Apps/Accounts/serializers.py
from rest_framework import serializers
from Apps.Accounts.models import User,Audit

class UserSerializer(serializers.ModelSerializer):
    password = serializers.CharField(write_only=True, required=True, style={'input_type': 'password'})
    confirm_password = serializers.CharField(write_only=True, required=True, style={'input_type': 'password'})
    
    class Meta:
        model = User
        fields = ['username', 'email', 'password', 'confirm_password', 'first_name', 'last_name']
        
    def validate(self, data):
        if data['password'] != data['confirm_password']:
            raise serializers.ValidationError("Las contraseñas no coinciden.")
        return data
    
    def create(self, validated_data):
        validated_data.pop('confirm_password')  # Eliminar confirmación de contraseña
        user = User.objects.create_user(**validated_data)
        return user

# Serializador para auditoría (solo lectura)
class AuditLogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Audit
        fields = ['created_at', 'updated_at', 'ip_address', 'user_agent']
        read_only_fields = fields
        